# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Family CFO API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions with filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: accountName
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: creditDebit
          in: query
          schema:
            type: string
            enum: [credit, debit]
        - name: isTransfer
          in: query
          schema:
            type: boolean
        - name: isRecurring
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
  /transactions/import:
    post:
      operationId: importTransactions
      tags: [transactions]
      summary: Import transactions from Frollo CSV (raw CSV content as text)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportTransactionsBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /transactions/{id}:
    patch:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update transaction (category, tags, isTransfer, isRecurring, userDescription)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionBody"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        transactionId:
          type: string
        description:
          type: string
        userDescription:
          type: string
          nullable: true
        amount:
          type: number
        currency:
          type: string
        transactionDate:
          type: string
        postedDate:
          type: string
          nullable: true
        accountNumber:
          type: string
        accountName:
          type: string
        creditDebit:
          type: string
          enum: [credit, debit]
        transactionType:
          type: string
        providerName:
          type: string
        merchantName:
          type: string
          nullable: true
        budgetCategory:
          type: string
          nullable: true
        categoryName:
          type: string
          nullable: true
        userCategory:
          type: string
          nullable: true
        userTags:
          type: array
          items:
            type: string
        notes:
          type: string
          nullable: true
        isTransfer:
          type: boolean
        isRecurring:
          type: boolean
        aiConfidenceScore:
          type: number
          nullable: true
        included:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - transactionId
        - description
        - amount
        - currency
        - transactionDate
        - accountNumber
        - accountName
        - creditDebit
        - transactionType
        - providerName
        - isTransfer
        - isRecurring
        - included
        - createdAt
        - updatedAt
        - userTags
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - transactions
        - total
        - page
        - limit
        - totalPages
    UpdateTransactionBody:
      type: object
      properties:
        userDescription:
          type: string
          nullable: true
        userCategory:
          type: string
          nullable: true
        userTags:
          type: array
          items:
            type: string
        notes:
          type: string
          nullable: true
        isTransfer:
          type: boolean
        isRecurring:
          type: boolean
        included:
          type: boolean
    ImportTransactionsBody:
      type: object
      properties:
        csvContent:
          type: string
          description: Raw CSV content from Frollo export
        fileName:
          type: string
      required:
        - csvContent
    ImportResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        updated:
          type: integer
        errors:
          type: integer
        message:
          type: string
      required:
        - imported
        - skipped
        - updated
        - errors
        - message
...